fix: cast Redis port/timeout to int/float for PHP 8.4 strict types
Redis::connect() in PHP 8.4 requires $port to be int and $timeout
to be float. When config values come from .mlc config files or env
vars, they arrive as strings, causing:

  Redis::connect(): Argument #2 ($port) must be of type int, string given

Explicit (int)/(float) casts fix the type mismatch.

+++ src/CacheManager.php
@@ -168,8 +168,8 @@ final class CacheManager
     {
         $redis   = new \Redis();
         $host    = $config['host'] ?? '127.0.0.1';
-        $port    = $config['port'] ?? 6379;
-        $timeout = $config['timeout'] ?? 0.0;
+        $port    = (int) ($config['port'] ?? 6379);
+        $timeout = (float) ($config['timeout'] ?? 0.0);
 
         $redis->connect($host, $port, $timeout);
